+++ lib/utils/slices/slices_test.go
@@ -19,6 +19,7 @@
 package slices
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
@@ -91,4 +92,93 @@ func TestFilterMapUnique(t *testing.T) {
 		expected := []string{"+a", "+b"}
 		require.Equal(t, expected, got)
 	})
+}
+
+// TestDuplicateKey tests slice deduplication via key function
+func TestDeduplicateKey(t *testing.T) {
+	t.Parallel()
+
+	stringTests := []struct {
+		name     string
+		slice    []string
+		keyFn    func(string) string
+		expected []string
+	}{
+		{
+			name:     "EmptyStringSlice",
+			slice:    []string{},
+			keyFn:    func(s string) string { return s },
+			expected: []string{},
+		},
+		{
+			name:     "NoStringDuplicates",
+			slice:    []string{"foo", "bar", "baz"},
+			keyFn:    func(s string) string { return s },
+			expected: []string{"foo", "bar", "baz"},
+		},
+		{
+			name:     "StringDuplicates",
+			slice:    []string{"foo", "bar", "bar", "bar", "baz", "baz"},
+			keyFn:    func(s string) string { return s },
+			expected: []string{"foo", "bar", "baz"},
+		},
+		{
+			name:     "StringDuplicatesWeirdKeyFn",
+			slice:    []string{"foo", "bar", "bar", "bar", "baz", "baz"},
+			keyFn:    func(s string) string { return "huh" },
+			expected: []string{"foo"},
+		},
+	}
+	for _, tt := range stringTests {
+		t.Run(tt.name, func(t *testing.T) {
+			res := DeduplicateKey(tt.slice, tt.keyFn)
+			require.Equal(t, tt.expected, res)
+		})
+	}
+
+	type dedupeStruct struct {
+		a string
+		b int
+		c bool
+	}
+	dedupeStructKeyFn := func(d dedupeStruct) string { return fmt.Sprintf("%s:%d:%v", d.a, d.b, d.c) }
+	structTests := []struct {
+		name     string
+		slice    []dedupeStruct
+		keyFn    func(d dedupeStruct) string
+		expected []dedupeStruct
+	}{
+		{
+			name:     "EmptySlice",
+			slice:    []dedupeStruct{},
+			keyFn:    dedupeStructKeyFn,
+			expected: []dedupeStruct{},
+		},
+		{
+			name: "NoStructDuplicates",
+			slice: []dedupeStruct{
+				{a: "foo", b: 1, c: true},
+				{a: "foo", b: 1, c: false},
+				{a: "foo", b: 2, c: true},
+				{a: "bar", b: 1, c: true},
+				{a: "bar", b: 1, c: false},
+				{a: "bar", b: 2, c: true},
+			},
+			keyFn: dedupeStructKeyFn,
+			expected: []dedupeStruct{
+				{a: "foo", b: 1, c: true},
+				{a: "foo", b: 1, c: false},
+				{a: "foo", b: 2, c: true},
+				{a: "bar", b: 1, c: true},
+				{a: "bar", b: 1, c: false},
+				{a: "bar", b: 2, c: true},
+			},
+		},
+	}
+	for _, tt := range structTests {
+		t.Run(tt.name, func(t *testing.T) {
+			res := DeduplicateKey(tt.slice, tt.keyFn)
+			require.Equal(t, tt.expected, res)
+		})
+	}
 }
